requirements.txt
psutil~=5.9.5
requests~=2.25.1
wmi
cryptography~=40.0.2
pillow~=9.5.0
pypiwin32
pycryptodome
alive-progress
PyAutoGUI~=0.9.53
pip
wheel
setuptools
sounddevice~=0.4.6
scipy~=1.10.1
keyboard~=0.13.5
winshell~=0.6
opencv-python~=4.7.0.72
numpy==1.23.5
pyperclip~=1.8.2
scapy~=2.5.0